+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## Unreleased
 
+- Support for Synthetic Run identification
+  ([#156](https://github.com/signalfx/splunk-otel-js/pull/156))
+
+## 0.10.0 (07-09-2021)
+
 - Rename `SPLUNK_SERVICE_NAME` to `OTEL_SERVICE_NAME`
+  ([#170](https://github.com/signalfx/splunk-otel-js/pull/170))
+- Upgrade to OpenTelemetry SDK 0.23.0
+  ([#173](https://github.com/signalfx/splunk-otel-js/pull/173))
 
 ## 0.9.0 (07-02-2021)